requirements.txt
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation 
opentelemetry-exporter-otlp